FLAS-40-Cards refactor: -replace header layout
